fix(review): move TestExecuteDispatchCancelledErrorLockRelease to test_api.py with strengthened lock assertion
Moved TestExecuteDispatchCancelledErrorLockRelease from test_fleet_semaphore.py
to test_api.py for cohesion (it tests execute_dispatch, not FleetSemaphore).

Added side-effect counter inside _raise_cancelled to prove the semaphore was
actually held when cancellation fired, distinguishing lock-acquired-then-released
from lock-never-acquired.
